replaced LookupReport with LookupReportDto for proper deserialization
