+++ da/src/main/java/org/jboss/bacon/da/rest/ReportsApi.java
@@ -1,4 +1,21 @@
-package org.jboss.bacon.da.rest.api;
+/*
+ * JBoss, Home of Professional Open Source.
+ * Copyright 2018 Red Hat, Inc., and individual contributors
+ * as indicated by the @author tags.
+ * <p>
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ * <p>
+ * http://www.apache.org/licenses/LICENSE-2.0
+ * <p>
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+package org.jboss.bacon.da.rest;
 
 import org.jboss.da.reports.model.request.AlignReportRequest;
 import org.jboss.da.reports.model.request.BuiltReportRequest;
@@ -8,7 +25,6 @@ import org.jboss.da.reports.model.request.SCMReportRequest;
 import org.jboss.da.reports.model.response.AdvancedReport;
 import org.jboss.da.reports.model.response.AlignReport;
 import org.jboss.da.reports.model.response.BuiltReport;
-import org.jboss.da.reports.model.response.LookupReport;
 import org.jboss.da.reports.model.response.Report;
 
 import javax.ws.rs.Consumes;
@@ -49,7 +65,7 @@ public interface ReportsApi {
     @Path("/lookup/gavs")
     @Consumes({ "application/json" })
     @Produces({ "application/json" })
-    List<LookupReport> lookupGav(LookupGAVsRequest gavRequest);
+    List<LookupReportDto> lookupGav(LookupGAVsRequest gavRequest);
 
     @POST
     @Path("/scm")
