--- da/src/main/java/org/jboss/bacon/da/rest/api/RestApplication.java
@@ -1,9 +1,0 @@
-package org.jboss.bacon.da.rest.api;
-
-import javax.ws.rs.ApplicationPath;
-import javax.ws.rs.core.Application;
-
-@ApplicationPath("/da/rest/v-1")
-public class RestApplication extends Application {
-
-}
