--- da/src/main/java/org/jboss/bacon/da/rest/package-info.java
@@ -1,31 +1,0 @@
-/*
- * JBoss, Home of Professional Open Source.
- * Copyright 2017 Red Hat, Inc., and individual contributors
- * as indicated by the @author tags.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- * http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-/**
- * TODO: This package contains classes generated from Swagger documentation of the DA server on April 24, 2019
- *
- * TODO: In some cases the generated classes were invalid and needed manual tweaking.
- *
- * TODO: I, the author, consider this classes a temporary work-around that should be replaced with either classes written from the scratch
- * TODO: or shared between PiG and DA.
- *
- *
- * TODO: Instructions for generating classes: https://github.com/OpenAPITools/openapi-generator/wiki/Server-stub-generator-HOWTO#java-jax-rs-resteasy
- *
- * TODO: At the very least, the classes that we don't use should be dropped from this package
- */
-package org.jboss.bacon.da.rest;
